UBER-911: Move search config
